pyproject.toml
[project]
name = "data-analyst-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.10",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
    "tabulate>=0.10.0",
]
